+++ src/main/java/org/jabref/gui/entryeditor/EntryEditor.java
@@ -11,8 +11,6 @@ import java.awt.event.ActionEvent;
 import java.awt.event.KeyAdapter;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -32,7 +30,10 @@ import javax.swing.JPopupMenu;
 import javax.swing.JToolBar;
 import javax.swing.SwingConstants;
 import javax.swing.SwingUtilities;
+import javax.swing.undo.UndoableEdit;
 
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.embed.swing.JFXPanel;
 import javafx.scene.Scene;
 import javafx.scene.control.Tab;
@@ -73,12 +74,13 @@ import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.search.SearchQueryHighlightListener;
 import org.jabref.logic.util.UpdateField;
 import org.jabref.model.EntryTypes;
-import org.jabref.model.FieldChange;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.EntryType;
+import org.jabref.model.entry.event.FieldAddedOrRemovedEvent;
 import org.jabref.preferences.JabRefPreferences;
 
+import com.google.common.eventbus.Subscribe;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.fxmisc.easybind.EasyBind;
@@ -96,6 +98,11 @@ import org.fxmisc.easybind.EasyBind;
 public class EntryEditor extends JPanel implements EntryContainer {
 
     private static final Log LOGGER = LogFactory.getLog(EntryEditor.class);
+
+    /**
+     * The default index number of the other fields tab
+     */
+    private static final int OTHER_FIELDS_DEFAULTPOSITION = 4;
 
     /**
      * A reference to the entry this object works on.
@@ -137,16 +144,21 @@ public class EntryEditor extends JPanel implements EntryContainer {
     private final UndoAction undoAction = new UndoAction();
     private final RedoAction redoAction = new RedoAction();
     private final List<SearchQueryHighlightListener> searchListeners = new ArrayList<>();
+    private final JFXPanel container;
 
     /**
      * Indicates that we are about to go to the next or previous entry
      */
-    private boolean movingToDifferentEntry;
+    private BooleanProperty movingToDifferentEntry = new SimpleBooleanProperty();
+    private EntryType entryType;
 
     public EntryEditor(JabRefFrame frame, BasePanel panel, BibEntry entry, String lastTabName) {
         this.frame = frame;
         this.panel = panel;
         this.entry = Objects.requireNonNull(entry);
+        entry.registerListener(this);
+        entryType = EntryTypes.getTypeOrDefault(entry.getType(),
+                this.frame.getCurrentBasePanel().getBibDatabaseContext().getMode());
 
         displayedBibEntryType = entry.getType();
 
@@ -156,7 +168,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
         setLayout(borderLayout);
         setupToolBar();
 
-        JFXPanel container = new JFXPanel();
+        container = new JFXPanel();
 
         container.addKeyListener(new KeyAdapter() {
 
@@ -187,6 +199,9 @@ public class EntryEditor extends JPanel implements EntryContainer {
         });
         DefaultTaskExecutor.runInJavaFXThread(() -> {
             addTabs(lastTabName);
+
+            tabbed.setStyle("-fx-font-size: " + Globals.prefs.getFontSizeFX() + "pt;");
+
             container.setScene(new Scene(tabbed));
         });
         add(container, BorderLayout.CENTER);
@@ -199,6 +214,57 @@ public class EntryEditor extends JPanel implements EntryContainer {
         });
 
         setupKeyBindings();
+    }
+
+    @Subscribe
+    public synchronized void listen(FieldAddedOrRemovedEvent event) {
+        // other field deleted -> update other fields tab
+        if (OtherFieldsTab.isOtherField(entryType, event.getFieldName())) {
+            DefaultTaskExecutor.runInJavaFXThread(() -> rebuildOtherFieldsTab());
+        }
+    }
+
+    private void rebuildOtherFieldsTab() {
+        int index = -1;
+        boolean isOtherFieldsTabSelected = false;
+
+        // find tab index and selection status
+        for (Tab tab: tabbed.getTabs()) {
+            if (tab instanceof OtherFieldsTab) {
+                index = tabbed.getTabs().indexOf(tab);
+                isOtherFieldsTabSelected = tabbed.getSelectionModel().isSelected(index);
+                break;
+            }
+        }
+
+        // rebuild tab at index and with prior selection status
+        if (index != -1) {
+            readdOtherFieldsTab(index, isOtherFieldsTabSelected);
+        } else {
+            // maybe the tab wasn't there but needs to be now
+            addNewOtherFieldsTabIfNeeded();
+        }
+    }
+
+    private void readdOtherFieldsTab(int index, boolean isOtherFieldsTabSelected) {
+        tabbed.getTabs().remove(index);
+        OtherFieldsTab tab = new OtherFieldsTab(frame, panel, entryType, this, entry);
+        // if there are no other fields left, no need to readd the tab
+        if (!(tab.getFields().size() == 0)) {
+            tabbed.getTabs().add(index, tab);
+        }
+        // select the new tab if it was selected before
+        if (isOtherFieldsTabSelected) {
+            tabbed.getSelectionModel().select(tab);
+        }
+    }
+
+    private void addNewOtherFieldsTabIfNeeded() {
+        OtherFieldsTab tab = new OtherFieldsTab(frame, panel, entryType, this, entry);
+        if (tab.getFields().size() > 0) {
+            // add it at default index, but that is just a guess
+            tabbed.getTabs().add(OTHER_FIELDS_DEFAULTPOSITION, tab);
+        }
     }
 
     private void selectLastUsedTab(String lastTabName) {
@@ -239,21 +305,19 @@ public class EntryEditor extends JPanel implements EntryContainer {
     }
 
     private void addTabs(String lastTabName) {
-        EntryType type = EntryTypes.getTypeOrDefault(entry.getType(),
-                this.frame.getCurrentBasePanel().getBibDatabaseContext().getMode());
 
         List<EntryEditorTab> tabs = new ArrayList<>();
 
         // Required fields
-        tabs.add(new RequiredFieldsTab(frame, panel, type, this, entry));
+        tabs.add(new RequiredFieldsTab(frame, panel, entryType, this, entry));
 
         // Optional fields
-        tabs.add(new OptionalFieldsTab(frame, panel, type, this, entry));
-        tabs.add(new OptionalFields2Tab(frame, panel, type, this, entry));
-        tabs.add(new DeprecatedFieldsTab(frame, panel, type, this, entry));
+        tabs.add(new OptionalFieldsTab(frame, panel, entryType, this, entry));
+        tabs.add(new OptionalFields2Tab(frame, panel, entryType, this, entry));
+        tabs.add(new DeprecatedFieldsTab(frame, panel, entryType, this, entry));
 
         // Other fields
-        tabs.add(new OtherFieldsTab(frame, panel, type, this, entry));
+        tabs.add(new OtherFieldsTab(frame, panel, entryType, this, entry));
 
         // General fields from preferences
         EntryEditorTabList tabList = Globals.prefs.getEntryEditorTabList();
@@ -270,7 +334,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
         tabs.add(new RelatedArticlesTab(entry));
 
         // Source tab
-        SourceTab sourceTab = new SourceTab(panel.getBibDatabaseContext(), entry);
+        SourceTab sourceTab = new SourceTab(panel, entry, movingToDifferentEntry);
         tabs.add(sourceTab);
 
         tabbed.getTabs().clear();
@@ -286,6 +350,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
         } else {
             selectLastUsedTab(lastTabName);
         }
+
     }
 
     public String getDisplayedBibEntryType() {
@@ -412,10 +477,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
 
     @Override
     public void requestFocus() {
-        EntryEditorTab activeTab = (EntryEditorTab) tabbed.getSelectionModel().getSelectedItem();
-        if (activeTab != null) {
-            activeTab.requestFocus();
-        }
+        container.requestFocus();
     }
 
     /**
@@ -457,7 +519,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
     }
 
     public void setMovingToDifferentEntry() {
-        movingToDifferentEntry = true;
+        movingToDifferentEntry.set(true);
         unregisterListeners();
     }
 
@@ -478,21 +540,6 @@ public class EntryEditor extends JPanel implements EntryContainer {
     private void warnEmptyBibtexkey() {
         panel.output(Localization.lang("Empty BibTeX key") + ". "
                 + Localization.lang("Grouping may not work for this entry."));
-    }
-
-    private boolean updateTimeStampIsSet() {
-        return Globals.prefs.getBoolean(JabRefPreferences.USE_TIME_STAMP)
-                && Globals.prefs.getBoolean(JabRefPreferences.UPDATE_TIMESTAMP);
-    }
-
-    /**
-     * Updates the timestamp of the given entry and returns the FieldChange
-     */
-    private Optional<FieldChange> doUpdateTimeStamp() {
-        String timeStampField = Globals.prefs.get(JabRefPreferences.TIME_STAMP_FIELD);
-        String timeStampFormat = Globals.prefs.get(JabRefPreferences.TIME_STAMP_FORMAT);
-        String timestamp = DateTimeFormatter.ofPattern(timeStampFormat).format(LocalDateTime.now());
-        return UpdateField.updateField(entry, timeStampField, timestamp);
     }
 
     private class TypeButton extends JButton {
@@ -590,8 +637,8 @@ public class EntryEditor extends JPanel implements EntryContainer {
 
         @Override
         public void actionPerformed(ActionEvent event) {
-            boolean movingAway = movingToDifferentEntry;
-            movingToDifferentEntry = false;
+            boolean movingAway = movingToDifferentEntry.get();
+            movingToDifferentEntry.set(false);
 
             if (event.getSource() instanceof TextField) {
                 // Storage from bibtex key field.
@@ -637,15 +684,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
 
                 // Add an UndoableKeyChange to the baseframe's undoManager.
                 UndoableKeyChange undoableKeyChange = new UndoableKeyChange(entry, oldValue, newValue);
-                if (updateTimeStampIsSet()) {
-                    NamedCompound ce = new NamedCompound(undoableKeyChange.getPresentationName());
-                    ce.addEdit(undoableKeyChange);
-                    doUpdateTimeStamp().ifPresent(fieldChange -> ce.addEdit(new UndoableFieldChange(fieldChange)));
-                    ce.end();
-                    panel.getUndoManager().addEdit(ce);
-                } else {
-                    panel.getUndoManager().addEdit(undoableKeyChange);
-                }
+                updateTimestamp(undoableKeyChange);
 
                 textField.setValidBackgroundColor();
 
@@ -707,18 +746,7 @@ public class EntryEditor extends JPanel implements EntryContainer {
                         // Add an UndoableFieldChange to the baseframe's undoManager.
                         UndoableFieldChange undoableFieldChange = new UndoableFieldChange(entry,
                                 fieldEditor.getFieldName(), oldValue, toSet);
-                        if (updateTimeStampIsSet()) {
-                            NamedCompound ce = new NamedCompound(undoableFieldChange.getPresentationName());
-                            ce.addEdit(undoableFieldChange);
-
-                            doUpdateTimeStamp()
-                                    .ifPresent(fieldChange -> ce.addEdit(new UndoableFieldChange(fieldChange)));
-                            ce.end();
-
-                            panel.getUndoManager().addEdit(ce);
-                        } else {
-                            panel.getUndoManager().addEdit(undoableFieldChange);
-                        }
+                        updateTimestamp(undoableFieldChange);
 
                         panel.markBaseChanged();
                     } catch (InvalidFieldValueException ex) {
@@ -743,6 +771,18 @@ public class EntryEditor extends JPanel implements EntryContainer {
                 SwingUtilities.invokeLater(() -> {
                     panel.getMainTable().ensureVisible(entry);
                 });
+            }
+        }
+
+        private void updateTimestamp(UndoableEdit undoableEdit) {
+            if (Globals.prefs.getTimestampPreferences().includeTimestamps()) {
+                NamedCompound compound = new NamedCompound(undoableEdit.getPresentationName());
+                compound.addEdit(undoableEdit);
+                UpdateField.updateField(entry, Globals.prefs.getTimestampPreferences().getTimestampField(), Globals.prefs.getTimestampPreferences().now()).ifPresent(fieldChange -> compound.addEdit(new UndoableFieldChange(fieldChange)));
+                compound.end();
+                panel.getUndoManager().addEdit(compound);
+            } else {
+                panel.getUndoManager().addEdit(undoableEdit);
             }
         }
     }