+++ src/main/java/org/jabref/gui/fieldeditors/MonthEditorViewModel.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javafx.util.StringConverter;
 
 import org.jabref.gui.autocompleter.AutoCompleteSuggestionProvider;
+import org.jabref.logic.integrity.FieldCheckers;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.Month;
 import org.jabref.model.strings.StringUtil;
@@ -13,8 +14,8 @@ import org.jabref.model.strings.StringUtil;
 public class MonthEditorViewModel extends OptionEditorViewModel<Month> {
     private BibDatabaseMode databaseMode;
 
-    public MonthEditorViewModel(String fieldName, AutoCompleteSuggestionProvider<?> suggestionProvider, BibDatabaseMode databaseMode) {
-        super(fieldName, suggestionProvider);
+    public MonthEditorViewModel(String fieldName, AutoCompleteSuggestionProvider<?> suggestionProvider, BibDatabaseMode databaseMode, FieldCheckers fieldCheckers) {
+        super(fieldName, suggestionProvider, fieldCheckers);
         this.databaseMode = databaseMode;
     }